Move inlay hints tests into implementation modules
